Version Bump and  ReadMe update
This change updates the readme to better reflect the usage and
requirements around using the rubypost helper scripts in addition to
aligning the manpage documentation.

Additionally, this adds the requirement of JQ over JSON_PP for pretty
printing JSON responses as JQ is a far more powerful and useful on its
own allowing for independent usage of JSON inspection by piping out the
result of `cat _response.json`. Future plans to incorporate this into
additional helper methods is on the roadmap.